[NCL-8159] Fix memory leak for metadata endpoint
We load all the log lines for a build in the metadata endpoint. By doing
so, we cause the entity manager to have a reference to all the LogLine
objects loaded in its first level cache. Unfortunately those LogLine
objects are big and as we load more of them, this causes an
OutOfMemoryError.

This commit tries to fix the memory leak by:
- Detaching the LogLine object from its entity manager to remove it from
  the first level cache and allowing the garbage collector to clean up
  after we have processed the LogLine
- removing the @Cacheable annotation for LogLine since we don't know the
  cache eviction policy