Rework on deletion of finallog
There seems to be a
[bug](https://hibernate.atlassian.net/browse/HHH-18967) in Hibernate 6
where the deletion of elements happen before deletion of the row in the
main table. Because of that, deletion of the row never happens since it
doesn't match the search query. I had to manually delete the rows
instead as a workaround.

The other change in this commit is that we are now using Postgresql in
our tests rather than H2. H2 exhibited weird issues with comparison for
timestamp. It didn't seem to be able to convert the timestamp with
timezone properly.